refactor: moved files & update base project files
