+++ package.json
@@ -36,7 +36,8 @@
   "scripts": {
     "start": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "prettier": "prettier --check ."
   },
   "eslintConfig": {
     "extends": [
@@ -62,6 +63,7 @@
     "@types/react-dom": "^18.0.0",
     "@vitejs/plugin-react": "^4.2.1",
     "miragejs": "^0.1.41",
+    "prettier": "3.5.0",
     "typescript": "^5.7.3"
   }
 }
