+++ src/components/data-table-column-header.tsx
@@ -2,14 +2,13 @@ import { Column } from '@tanstack/react-table';
 import { ArrowDown, ArrowUp, ChevronsUpDown, EyeOff } from 'lucide-react';
 
 import { cn } from '@/lib/utils';
-import { Button } from './button';
+import { Button } from './shadcn/button';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from './dropdown-menu';
+} from './shadcn/dropdown-menu';
 import React from 'react';
 
 interface DataTableColumnHeaderProps<TData, TValue>
