+++ src/components/shadcn/columns.tsx
@@ -1,5 +1,5 @@
 import { ColumnDef } from '@tanstack/react-table';
-import { DataTableColumnHeader } from './data-table-column-header';
+import { DataTableColumnHeader } from '../data-table-column-header';
 import React from 'react';
 import { Author, Category, Post } from '@/lib/schema';
